\name{irismod}
\docType{data}
\alias{irismod}
\title{A Slight Modification to Edgar Anderson's Iris Data}
\description{
The famous iris data set, but with the species column altered from factor to binary and renamed \code{Is.Virginica}. 1's represent the Virginica Species whereas the (-1)'s represent "veriscolor" or "virginica".
}
\usage{irismod}
\format{A dataframe}
\source{Fischer, R. A. (1935) The use of multiple measurements in taxonomic problems. \emph{Annals of Eugenics}, \bold{7}, Part II, 179--188}

\keyword{datasets}
