\name{RSofia-internal}
\alias{RSofiaFacade}
\alias{sofia_ml_test}
\alias{sofia_ml_train}

\title{Internal RSofia Functions} 
\description{
  Internal RSofia functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{ internal }
